pyproject.toml
[tool.poetry]
name = "python-books"
version = "0.1.0"
description = ""
authors = ["Matvey <matveykawork@gmail.com>"]
readme = "README.md"
packages = [{include = "python_books"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.2"
aiosqlite = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
